codenext_blog/requirements.txt
asgiref==3.8.1
autopep8==2.3.1
crispy-bootstrap5==2024.10
Django==4.2
django-crispy-forms==2.3
django-silk==5.3.1
djangorestframework==3.15.2
gprof2dot==2024.6.6
pillow==11.0.0
psycopg2-binary==2.9.10
pycodestyle==2.12.1
python-dotenv==1.0.1
sqlparse==0.5.1
